feat: Include host name in email subject
If enrollment reports are being generated on multiple systems, it
helps to be able to tell them apart by source.

Thus, include a configurable host name in the email subject. This
defaults to ansible_host, but when used in conjunction with
tutor-contrib-enrollmentreports it should be populated with the
LMS_HOST variable.