feat: Make the enrollment report frequency configurable
Previously, enrollment reports would only ever be generated monthly.

With this change, we can set a new variable,
"enrollment_report_frequency", which can be "monthly" or "weekly".

If set to "monthly" (the default), the playbook retains the previous
behaviour of generating a report covering enrollments from the past
month. If set to "weekly", it generates a weekly report instead.